fix(desktop): scope v2 preset seed marker to the active organization
The v2TerminalPresets collection is per-organization but the seed
marker was global, so users switching orgs saw an empty preset bar
because the flag claimed they were already seeded. Moves the shared
hook to _authenticated/hooks/useSeedDefaultV2Presets, keys the
localStorage marker on the active organization id, and migrates any
pre-fix global marker forward to the currently-active org on first
run so legacy users don't re-seed into duplicates.

Also consolidates V2PresetsSection's inline seed into the shared
hook, and removes the vestigial canOpen prop from V2PresetBarItem
(V2PresetsBar requires a workspaceId, so canOpen was always true).
